backend/requirements.txt
# aiohappyeyeballs==2.4.6
# aiohttp==3.11.12
# aiosignal==1.3.2
# annotated-types==0.7.0
# anyio==4.8.0
# asgiref==3.8.1
# asttokens==3.0.0
async-timeout==4.0.3
attrs==25.1.0
# backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.13.3
build==1.2.2.post1
cachetools==5.5.1
certifi==2025.1.31
# charset-normalizer==3.4.1
# chroma-hnswlib==0.7.6
# chromadb==0.6.3
click==8.1.8
# colorama==0.4.6
# coloredlogs==15.0.1
# comm==0.2.2
# contourpy==1.3.1
# cycler==0.12.1
# dataclasses-json==0.6.7
# debugpy==1.8.12
# decorator==5.1.1
# Deprecated==1.2.18
# distro==1.9.0
# duckduckgo_search==7.5.2
# durationpy==0.9
# exceptiongroup==1.2.2
# executing==2.2.0
# faiss-cpu==1.10.0
fastapi==0.115.8
# fastjsonschema==2.21.1
# filelock==3.17.0
# filetype==1.2.0
# flatbuffers==25.2.10
# fonttools==4.56.0
# frozendict==2.4.6
# frozenlist==1.5.0
# fsspec==2025.2.0
# google-ai-generativelanguage==0.6.15
# google-api-core==2.24.1
# google-api-python-client==2.160.0
# google-auth==2.38.0
# google-auth-httplib2==0.2.0
# google-generativeai==0.8.4
# google_search_results==2.4.2
# googleapis-common-protos==1.67.0rc1
# greenlet==3.1.1
# groq==0.18.0
# grpcio==1.70.0
# grpcio-status==1.70.0
# h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
# humanfriendly==10.0
# idna==3.10
# importlib_metadata==8.5.0
# importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.32.0
# jedi==0.19.2
# Jinja2==3.1.6
# jiter==0.9.0
# joblib==1.4.2
# jsonlines==4.0.0
# jsonpatch==1.33
# jsonpointer==3.0.0
# jsonschema==4.23.0
# jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
# kiwisolver==1.4.8
# kubernetes==32.0.0
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.45
langchain-experimental==0.3.4
langchain-openai==0.3.9
langchain-text-splitters==0.3.6
langchainhub==0.1.21
langgraph==0.2.70
langgraph-prebuilt
langgraph-checkpoint==2.0.12
langgraph-sdk==0.1.51
langsmith==0.3.8
# loguru==0.7.3
# lxml==5.3.1
# markdown-it-py==3.0.0
# MarkupSafe==2.1.1
# marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp-use
# mdurl==0.1.2
# mmh3==5.1.0 
# monotonic==1.6
# mpmath==1.3.0
# msgpack==1.1.0
# multidict==6.1.0
# multitasking==0.0.11
# mypy-extensions==1.0.0
# nbformat==5.10.4
# nest-asyncio==1.6.0
# networkx==3.4.2
# nomic==3.4.1
numpy==1.26.4
oauthlib==3.2.2
# onnxruntime==1.20.1
openai==1.66.5
# opentelemetry-api==1.30.0
# opentelemetry-exporter-otlp-proto-common==1.30.0
# opentelemetry-exporter-otlp-proto-grpc==1.30.0
# opentelemetry-instrumentation==0.51b0
# opentelemetry-instrumentation-asgi==0.51b0
# opentelemetry-instrumentation-fastapi==0.51b0
# opentelemetry-proto==1.30.0
# opentelemetry-sdk==1.30.0
# opentelemetry-semantic-conventions==0.51b0
# opentelemetry-util-http==0.51b0
# orjson==3.10.15
# overrides==7.7.0
packaging==24.2
pandas==2.2.3
# parso==0.8.4
# peewee==3.17.9
# pillow==10.4.0
# platformdirs==4.3.6
# posthog==3.12.0
# primp==0.14.0
# prompt_toolkit==3.0.50
# propcache==0.2.1
# proto-plus==1.26.0
# protobuf==5.29.3
# psutil==6.1.1
psycopg2-binary
# pure_eval==0.2.3
pyarrow==19.0.0
# pyasn1==0.6.1
# pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
# Pygments==2.19.1
# PyJWT==2.10.1
# pyparsing==3.2.1
# PyPika==0.48.9
# pyproject_hooks==1.2.0
# pyreadline3==3.5.4
# python-dateutil==2.9.0.post0
python-dotenv==1.0.1
# pytz==2025.1
# pywin32==308
# PyYAML==6.0.2
# pyzmq==26.2.1
# referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
# rpds-py==0.24.0
# rsa==4.9
# safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.4.1
# shellingham==1.5.4
# six==1.17.0
# sniffio==1.3.1
# soupsieve==2.6
SQLAlchemy==2.0.38
sqlmodel
stack-data==0.6.3
starlette==0.45.3
# sympy==1.13.1
tavily-python==0.5.1
# tenacity==9.0.0
# threadpoolctl==3.5.0
# tiktoken==0.8.0
tokenizers==0.21.0
tomli==2.2.1
torch==2.6.0
tornado==6.4.2
# tqdm==4.67.1
# traitlets==5.14.3
# transformers==4.48.3
# typer==0.15.1
# types-requests==2.32.0.20241016
# typing-inspect==0.9.0
# typing_extensions==4.12.2
# tzdata==2025.1
# uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
# watchfiles==1.0.4
# wcwidth==0.2.13
# webencodings==0.5.1
# websocket-client==1.8.0
# websockets==14.2
# wikipedia==1.4.0
# win32_setctime==1.2.0
# wrapt==1.17.2
# yarl==1.18.3
# yfinance==0.2.52
# youtube-search==2.1.2
# zipp==3.21.0
# zstandard==0.23.0
# fastapi[all]
streamlit
langchain-community
-e .